chore: Update UI plugins and add new utility plugins
This commit updates the configuration files for UI plugins and adds new utility plugins. The `alpha.nix` file was added to the `plugins/ui` directory, and the `lualine.nix` file was modified to disable the `statusline` for the `startup` filetype. Additionally, the `zellij.nix` and `telescope.nix` files were added to the `plugins/utils` directory.