chore: Update indent-blankline, navic, and default.nix configurations
This commit updates the configurations for the `indent-blankline`, `navic`, and `default.nix` files.

In the `indent-blankline.nix` file, the `enable` property is moved inside the `indent-blankline` block.

In the `navic.nix` file, the `click`, `depthLimit`, `depthlimitdepthLimitIndicator`, `highlight`, `lazyUpdateContext`, `safeOutput`, and `separator` properties are modified.

In the `default.nix` file, the `dim_inactive` property is modified.

These changes improve the functionality and appearance of the editor plugins and themes.